Add ``get_setting`` template tag.
The template tag is available in the ``core_settings`` module
and has been added as an accessory to read the app settings
from the templates.

This is in order to avoid to modify an existing view only
to add a setting in the template context.

e.g. Change password view.

Usage:

    {% load core_settings %}
    {% get_setting 'SITE_URL' %}