Add form to import users in batches.
https://github.com/madewithbytes/us_ignite/issues/18

The form receives a payload with the ``user`` field
listing a new user on each line with the following format::

    name, email

The form validates every entry and removes any blank lines.

During the ``clean_users`` method the form will transform
the payload in a list of tuples containing the name and email.

Refers #18.