Add decorator to restrict access by group to certain areas.
https://github.com/madewithbytes/us_ignite/issues/2

The decorator allows access to restricted views to
authenticated users that belon to the specified groups.

The decorator takes a group or a list of groups,
and then validates if the user is part of that group.

If the user is not authenticated redirects it to the
login page.

If the user is authenticated but has not enough access
shows an unavailable page.

Closes #2.