Send welcome email when user activates account.
https://github.com/madewithbytes/us_ignite/issues/1

When a new user is created via Mozilla Persona, or
activated using the US Ignite registration mechanics
a welcome email is sent.

The welcome email is sent using: ``browserid.user_created`` signal
and ``registration.user_activated`` signals, because when
these signals are emitted the account is ready to be
used.

Closes #1.