Replaced ``render`` with ``TemplateResponse``.
By switching these components it is easier to unittest
the views since the response from the view can be inspected.

Oherwise the only way to inspect the response is creating
integration tests using Django test client, since render
only returns an ``HTTPResponse`` which only carries the
``HTML`` output of the view.

As a consequence of this several integration tests were
translated to unittest.