List social links in the ``Application`` profile.
https://github.com/madewithbytes/us_ignite/issues/20

The application can have different social media accounts
associated with the profile.

These links are detected and rendered accordingly with
a ``templatetag``.

If the link is not supported or failed  an empty string
is returned.

The idea is to render some extra information from the
given URLS and show the progress of the application.

Closes #20.