Add ``MembershipForm`` to validate collaborators for an app.
https://github.com/madewithbytes/us_ignite/issues/74

The form receives a list of email addresses separated
by a new line.

The email address is validated with the registered users,
since being a member of an application gives the account
extra permissions for the application.

The ``clean_collaborators`` methods transform the payload
in a list of registered users. If an invalid email address
or a non-registered user is found, the form returns an
error.