Add mechanics for ``Users`` to associate with a community.
https://github.com/madewithbytes/us_ignite/issues/98

The process is initiated from the ``Hub`` page, since
makes more for the user to subscribe to a commiunity
after the user has seen its content.

The view to add the membership is only available
via ``POST`` initiated from the ``Hub`` detail page.

The hub detail page determines when a user is the
guardian or a member of this community.