Implement Ignite community approval flow.
https://github.com/madewithbytes/us_ignite/issues/101

Once the user has sent an request to be an Ignite Community
the administrators will receive a notification of the request.

The request will be resolved in the admin section, on the top
right corner there is a link got resolve the request.

The form receives only two values: ``status`` and ``notes``.

- ``status`` determines the resolution of the request
- ``notes`` an optional field that will be feedback to the
user that generated the request.

If the request has been marked as ``approved`` the view
will generate a ``Hub`` where the ``user is the ``guardian``
and associate the request with it.