Add app to generate dummy content.
https://github.com/madewithbytes/us_ignite/issues/73

The application provides a ``dummy_generate_content`` command
to populate the models by generating phrases, paragraphs and
images.

By default the command is dissabled but can be enabled by
adding ``us_ignite.dummy`` to the ``INSTALLED_APPS`` setting.

Requires a file with a list of words, the path to the file
is specified in ``WORDS_PATH``