Add view to enter an application into a challenge.
https://github.com/madewithbytes/us_ignite/issues/111

Once the challenge is open a published application
can be entered by the ``owner`` of the app.

The entry has a ``draft`` a ``submitted`` status

The challenge form is generated dynamically from the
questions added in the admin, these questions can be
compulsory or not.

If the form is marked as draft the user can save
the progress and come back to review the entry
at a later point.