Add detail page for a challenge.
https://github.com/madewithbytes/us_ignite/issues/113

Users should be able to see a challenge detail page when
the challenge is published.

Registered users will see a list of owned applications
to allow them to participate in the Challenge when it
is open.

The applications and their entries are matched in a
manager and bundled in a ``AppEntry`` named tuple.