Add redirect handler for POST legacy URLS.
The ``Post`` legacy URLS have the following format:

http://us-ignite.org/2013/04/building-a-cathedral-dont-lose-sight-of-the-bricks/

The new URL format has a ``/blog/`` segment in the URL to
separate it from the rest of the content.

http://us-ignite.org/blog/2013/04/building-a-cathedral-dont-lose-sight-of-the-bricks/

The URL is processed in a similar fashion as the original
imported content in order to accurately match an existing
piece of content.