Use the form timezone to localize datetime in events.
https://github.com/madewithbytes/us_ignite/issues/287

The event form replaces the timezone of the start and
end datetime objects, using the selected timezone.

The timezone is replaced by converting the objects to
native dates and then making them back aware with the
new timezone.

The form localises the datetime fields using the
event timezone before the form is rendered.

These two measures make sure that the editing and
saving the datetime objects preserve the right
timezone.